Добавлен глобальный обработчик исключений и улучшена проверка безопасности

+++ src/main/java/com/titov/securedforumapi/services/PostService.java
@@ -5,13 +5,17 @@ import com.titov.securedforumapi.dto.request.AddPostRequest;
 import com.titov.securedforumapi.dto.response.ListPostsResponse;
 import com.titov.securedforumapi.dto.response.PostResponse;
 import com.titov.securedforumapi.models.Post;
+import com.titov.securedforumapi.models.Role;
 import com.titov.securedforumapi.models.User;
 import com.titov.securedforumapi.repositories.PostRepo;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+
 
 @Service
 @AllArgsConstructor
@@ -39,12 +43,20 @@ public class PostService {
 
         Post post = postRepo.findPostById(id);
 
+        if (post == null) {
+            throw new EntityNotFoundException("Пост с ID " + id + " не найден");
+        }
+
         post.setTitle(newPost.title());
         post.setContent(newPost.content());
         return mapper.postToPostResponse(postRepo.save(post));
     }
 
     public void deletePostById(long id) {
+        User user = postRepo.findPostById(id).getUser();
+        if ((findCurrentUser() != user) || !hasRoleAdmin(user)) {
+            throw new AccessDeniedException("У вас нет прав для выполнения этого действия");
+        }
         postRepo.deleteById(id);
     }
 
@@ -53,5 +65,10 @@ public class PostService {
         UserDetails userDetails = (UserDetails) authentication.getPrincipal();
         String username = userDetails.getUsername();
         return userService.findByUsername(username);
+    }
+
+    private boolean hasRoleAdmin(User user) {
+        return user.getRoles().stream()
+                .anyMatch(role -> role.getName().equals("ADMIN"));
     }
 }
